requirements.txt
pandas>=1.3.0
numpy>=1.20.0
statsmodels>=0.13.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
networkx>=2.6.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.4.0